- Implemented architecture to easily make GameController sibling systems accessible
Add a "GameControllerSystem" class that the other systems can inherit
off of. GameController exclusively manages the singleton behaviour of it
host game object, so the other components in the GameController object
should not be doing that, but they are also singleton in nature.
GameControllerSystem therefore only does some templated static instance
caching so that the rest of the game can freely access and use the
systems in the GameController.