- Improved projectile launcher architecture
Was able to get rid of a few subclasses by making the abstract
projectile launcher more flexible. Also added an ID system to the
launchers to accommodate game objects having more than one. They can now
specify a launcher ID in the message and only the launcher with the
matching ID will fire.